Updates node interface, adds missing inbound abstraction
