+++ node/src/client/mod.rs
@@ -24,18 +24,7 @@ use snarkos_node_tcp::{
     protocols::{Disconnect, Handshake, Reading, Writing},
     P2P,
 };
-use snarkvm::prelude::{
-    Address,
-    Block,
-    CoinbasePuzzle,
-    ConsensusStorage,
-    EpochChallenge,
-    Header,
-    Network,
-    PrivateKey,
-    ProverSolution,
-    ViewKey,
-};
+use snarkvm::prelude::{Block, CoinbasePuzzle, ConsensusStorage, EpochChallenge, Header, Network, ProverSolution};
 
 use anyhow::Result;
 use core::marker::PhantomData;
@@ -45,8 +34,6 @@ use std::{net::SocketAddr, sync::Arc};
 /// A client node is a full node, capable of querying with the network.
 #[derive(Clone)]
 pub struct Client<N: Network, C: ConsensusStorage<N>> {
-    /// The account of the node.
-    account: Account<N>,
     /// The router of the node.
     router: Router<N>,
     /// The genesis block.
@@ -74,7 +61,7 @@ impl<N: Network, C: ConsensusStorage<N>> Client<N, C> {
         let router = Router::new(
             node_ip,
             NodeType::Client,
-            account.clone(),
+            account,
             trusted_peers,
             Self::MAXIMUM_NUMBER_OF_PEERS as u16,
             dev.is_some(),
@@ -84,7 +71,6 @@ impl<N: Network, C: ConsensusStorage<N>> Client<N, C> {
         let coinbase_puzzle = CoinbasePuzzle::<N>::load()?;
         // Initialize the node.
         let node = Self {
-            account,
             router,
             genesis,
             coinbase_puzzle,
@@ -103,31 +89,6 @@ impl<N: Network, C: ConsensusStorage<N>> Client<N, C> {
 
 #[async_trait]
 impl<N: Network, C: ConsensusStorage<N>> NodeInterface<N> for Client<N, C> {
-    /// Returns the node type.
-    fn node_type(&self) -> NodeType {
-        self.router.node_type()
-    }
-
-    /// Returns the account private key of the node.
-    fn private_key(&self) -> &PrivateKey<N> {
-        self.account.private_key()
-    }
-
-    /// Returns the account view key of the node.
-    fn view_key(&self) -> &ViewKey<N> {
-        self.account.view_key()
-    }
-
-    /// Returns the account address of the node.
-    fn address(&self) -> Address<N> {
-        self.account.address()
-    }
-
-    /// Returns `true` if the node is in development mode.
-    fn is_dev(&self) -> bool {
-        self.router.is_dev()
-    }
-
     /// Shuts down the node.
     async fn shut_down(&self) {
         info!("Shutting down...");
