+++ node/src/traits.rs
@@ -21,19 +21,29 @@ use snarkvm::prelude::{Address, Network, PrivateKey, ViewKey};
 #[async_trait]
 pub trait NodeInterface<N: Network>: Routing<N> {
     /// Returns the node type.
-    fn node_type(&self) -> NodeType;
+    fn node_type(&self) -> NodeType {
+        self.router().node_type()
+    }
 
     /// Returns the account private key of the node.
-    fn private_key(&self) -> &PrivateKey<N>;
+    fn private_key(&self) -> &PrivateKey<N> {
+        self.router().private_key()
+    }
 
     /// Returns the account view key of the node.
-    fn view_key(&self) -> &ViewKey<N>;
+    fn view_key(&self) -> &ViewKey<N> {
+        self.router().view_key()
+    }
 
     /// Returns the account address of the node.
-    fn address(&self) -> Address<N>;
+    fn address(&self) -> Address<N> {
+        self.router().address()
+    }
 
     /// Returns `true` if the node is in development mode.
-    fn is_dev(&self) -> bool;
+    fn is_dev(&self) -> bool {
+        self.router().is_dev()
+    }
 
     /// Handles OS signals for the node to intercept and perform a clean shutdown.
     /// Note: Only Ctrl-C is supported; it should work on both Unix-family systems and Windows.
